+++ client/Nav/navbar.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {connect} from 'react-redux'
-import {Link} from 'react-router-dom'
-import {logout} from '../store'
+import { connect } from 'react-redux'
+import { Link } from 'react-router-dom'
+import { logout } from '../allDispatch'
 
-const Navbar = ({handleClick, isLoggedIn}) => (
+const Navbar = ({ handleClick, isLoggedIn }) => (
   <div>
     <h1>Welcome to Graced Pickles</h1>
     #blessed
@@ -19,14 +19,14 @@ const Navbar = ({handleClick, isLoggedIn}) => (
           <Link to="/pickles">View All Pickles</Link>
         </div>
       ) : (
-        <div>
-          {/* The navbar will show these links before you log in */}
-          <Link to="/login">Login</Link>
-          <Link to="/signup">Sign Up</Link>
-          <Link to="/pickles">View All Pickles</Link>
-          <Link to="/cart">View Your Shopping Cart</Link>
-        </div>
-      )}
+          <div>
+            {/* The navbar will show these links before you log in */}
+            <Link to="/login">Login</Link>
+            <Link to="/signup">Sign Up</Link>
+            <Link to="/pickles">View All Pickles</Link>
+            <Link to="/cart">View Your Shopping Cart</Link>
+          </div>
+        )}
     </nav>
     <hr />
   </div>
