+++ client/User/Home/user-home.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {connect} from 'react-redux'
+import { connect } from 'react-redux'
 
 /**
  * COMPONENT
  */
 export const UserHome = props => {
-  const {email} = props
+  const { email } = props
 
   return (
     <div>
