+++ client/allComponents.js
@@ -1,0 +1,12 @@
+/**
+ * `components/index.js` exists simply as a 'central export' for our components.
+ * This way, we can import all of our components from the same place, rather than
+ * having to figure out which file they belong to!
+ */
+export { default as Navbar } from './Nav/navbar'
+export { default as UserHome } from './User/Home/user-home'
+export { Login, Signup } from './Auth/auth-form'
+export { default as AllPickles } from './Pickles/All/component'
+export { default as SinglePickle } from './Pickles/Single/component'
+export { default as Cart } from './Cart/component'
+
