+++ client/allDispatch.js
@@ -1,0 +1,7 @@
+// this way we do not care where the redux is defined
+
+
+export { me, auth, logout, fetchUserCart, userAddsToCart } from './User/redux'
+export { addToCart, removeOnePickleFromCart, removeFromCart } from './Cart/redux'
+export { fetchPickles } from './Pickles/All/redux'
+export { resetPickles, fetchSinglePickle } from './Pickles/Single/redux'
