+++ client/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import {Provider} from 'react-redux'
-import {Router} from 'react-router-dom'
+import { Provider } from 'react-redux'
+import { Router } from 'react-router-dom'
 import history from './history'
-import store from './store'
+import store from './allRedux'
 import App from './app'
 
 // establishes socket connection
